+++ src/main/java/org/pharmgkb/pharmcat/definition/model/VariantLocus.java
@@ -68,7 +68,7 @@ public class VariantLocus implements Comparable<VariantLocus> {
   /**
    * Gets the VCF position for this variant.
    */
-  public int getVcfPosition() {
+  public long getVcfPosition() {
     if (m_type == VariantType.DEL) {
       return m_position - 1;
     }
